requirements.txt
# Telegram Bot Framework (асинхронный)
aiogram==3.13.1

# RSS парсинг
feedparser==6.0.11

# HTML парсинг для извлечения полного текста статей
beautifulsoup4==4.12.3
lxml==5.3.0
trafilatura==1.12.2

# Асинхронные HTTP запросы
aiohttp==3.10.10

# Парсинг дат из RSS
python-dateutil==2.9.0.post0

# Работа с таймзонами
pytz==2024.2

# Логирование (улучшенное, опционально)
colorlog==6.9.0

# Загрузка переменных окружения
python-dotenv==1.0.1

# SQLite уже встроен в Python, дополнительная библиотека не требуется
# hashlib также встроен в Python (используется для хеширования контента)

